pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [ "setuptools>=64", "setuptools-scm>=8" ]

[project]
name = "scitrail"
version = "0.0.1"
description = "Analyze scientific literature with structured criteria to reveal the current state of research."
readme = "README.md"
authors = [
  { name = "Dave Bunten" },
]
requires-python = ">=3.11"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "fire>=0.7.0",
  "instructor>=1.7.0",
  "pyalex>=0.19",
  "pydantic>=2.10",
  "pyyaml>=6.0.2",
]
scripts.scitrail = "scitrail.cli:trigger"

[project.optional-dependencies]
local-llm = [
  "llama-cpp-python>=0.2.90",
]

[dependency-groups]
dev = [
  "poethepoet>=0.34",
  "pytest>=8.3.5",
]
docs = [
  "myst-nb>=1.2",
  "pydata-sphinx-theme>=0.16.1",
  "sphinx>=8.2.3",
]
notebooks = [
  "black>=25.1",
  "isort>=6.0.1",
  "jupyterlab>=4.4.2",
  "jupyterlab-code-formatter>=3.0.2",
  "jupytext>=1.17.1",
]

[tool.setuptools_scm]
root = "."
version_scheme = "no-guess-dev"
local_scheme = "no-local-version"
version_file = "src/scitrail/_version.py"

[tool.ruff]
target-version = "py311"
line-length = 88
fix = true
lint.select = [
  # flake8-builtins
  "A",
  # flake8-annotations
  "ANN",
  # flake8-comprehensions
  "C4",
  # mccabe
  "C90",
  # pycodestyle
  "E",
  # pyflakes
  "F",
  # isort
  "I",
  # pylint
  "PL",
  # ruff
  "RUF",
  # flake8-simplify
  "SIM",
  "W",
]
# Ignore `F401` (unused imports) for `__init__.py` file
lint.per-file-ignores."__init__.py" = [ "F401" ]
lint.per-file-ignores."tests/*.py" = [ "ANN001", "PLR2004" ]

[tool.pytest.ini_options]
pythonpath = [ "src" ]

[tool.poe.tasks]
# tasks for development environments
test.shell = """
pytest
"""
jupyter.shell = """
jupyter lab
"""
docs-notebook-cu-quantum.shell = """
uv run --group docs --group notebooks --frozen jupyter nbconvert --to notebook --execute --inplace --ExecutePreprocessor.timeout=1200 docs/src/examples/cu_anschutz_quantum.ipynb
"""

[tool.jupytext]
formats = "ipynb,py:light"

[tool.vulture]
min_confidence = 90
paths = [ "src/scitrail", "tests" ]
